Séparation de ertain morceaux de la carte dands un autre fichier.
